WGRND-639: update tests
