+++ .github/workflows/node.js.yml
@@ -29,12 +29,8 @@ jobs:
       - run: npm run start -d & sleep 40
       - name: Cypress authorization tests
         run: npm run cy:run:auth
-      - name: Cypress run navigation tests
-        run: npm run cy:run:nav
-      - name: Cypress CRU tests
-        run: npm run cy:run:cru
-      - name: Cypress run permissions tests
-        run: npm run cy:run:per
+      - name: Cypress run
+        run: npm run cypress:run:chrome
       #         uses: cypress-io/github-action@v2
 #         with:
 #           wait-on-timeout: 80
