+++ cypress.json
@@ -26,6 +26,7 @@
   "defaultCommandTimeout": 12000,
   "requestTimeout": 10000,
   "video": false,
+  "videoCompression": 0,
   "viewportWidth": 1920,
   "viewportHeight": 1080,
   "reporter": "mochawesome",
