+++ cypress/integration/common/login.js
@@ -70,6 +70,10 @@ Then(/^Login as new user 2 without UI$/, () => {
         },
       }).as('Set user2 token')
   })
+  cy.wait('@getFolder').then((xhr) => {
+    expect(xhr.responseBody).to.not.have.property('stack')
+    console.log(xhr.responseBody)
+  })
 });
 
 When(/^Login as new user 3 without UI$/, () => {
