+++ cypress/integration/common/ownerships.js
@@ -43,7 +43,7 @@ Then(/^"([^"]*)" option from pop-up window is not visible$/,  () => {
 Given(/^Choose the "([^"]*)" option from pop-up window$/, (option) => {
   cy.get('.ant-modal-header').should('be.visible')
   cy.get('#form_in_modal_permissions').should('be.visible').click().wait(1000)
-  cy.contains(option).click(1000)
+  cy.contains(option).click().wait(1000)
 });
 
 Given(/^Press "([^"]*)"$/, (button) => {
