+++ package.json
@@ -97,7 +97,7 @@
     "cy:run:auth": "cypress run --browser chrome --spec \"cypress/integration/auth/*.feature\"",
     "cy:run:cru": "cypress run --browser chrome --spec \"cypress/integration/CRU/*.feature\"",
     "cy:run:nav": "cypress run --browser chrome --spec \"cypress/integration/navigation/*.feature\"",
-    "cy:run:per": "cypress run --browser chrome --spec \"cypress/integration/permissions/*.feature\"",
+    "cy:run:per": "cypress run --browser chrome --spec \"cypress/integration/permissions/1_file_ownership.feature\"",
     "cypress:run:chrome": "cypress run --browser chrome --spec \"cypress/integration/*/*.feature\"",
     "cypress:tags": "cypress run -e TAGS='@step'",
     "cypress:report": "cypress run --spec cypress/integration/API/POST_user.feature --reporter mochawesome && open ./mochawesome-report/mochawesome.html",
